+++ .idea/misc.xml
@@ -4,7 +4,7 @@
     <option name="filePathToZoomLevelMap">
       <map>
         <entry key="..\:/Users/alex2/Desktop/App_IUDigital/app/src/main/res/layout/activity_main.xml" value="0.18541666666666667" />
-        <entry key="..\:/Users/alex2/Desktop/App_IUDigital/app/src/main/res/layout/activity_registro_usuario.xml" value="0.18541666666666667" />
+        <entry key="..\:/Users/alex2/Desktop/App_IUDigital/app/src/main/res/layout/activity_registro_usuario.xml" value="0.4" />
       </map>
     </option>
   </component>
