+++ app/src/main/java/com/example/app_iudigital/registroUsuario.java
@@ -19,7 +19,7 @@ public class registroUsuario extends AppCompatActivity {
         volver.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                Intent intent = new Intent (v.getContext(), MainActivity.class);
+                Intent intent = new Intent (v.getContext(), LoginActivity.class);
                 startActivityForResult(intent, 0);
             }
         });
